Added dialogs for rewards, badges, experiences, and majors
