+++ src/router/router.js
@@ -2,21 +2,15 @@ import { createRouter, createWebHistory } from "vue-router";
 import Login from "../views/Login.vue";
 import TasksPage from "../views/admin/task/TasksPage.vue";
 import EventCardPage from "../views/admin/event/EventsPage.vue";
-//import EventAddEditPage from "../views/admin/event/EventAddEditDialog.vue";
 import { userStore } from "../stores/userStore";
 import NotFound from "../views/NotFound.vue";
 import Unauthorized from "../views/Unauthorized.vue";
-//import TaskAddEditPage from "../views/admin/task/TaskAddEditDialog.vue";
 import BadgesPage from "../views/admin/badge/BadgesPage.vue";
-import BadgeAddEditPage from "../views/admin/badge/BadgeAddEditPage.vue";
-import ExperienceAddEditPage from "../views/admin/experience/ExperienceAddEditPage.vue";
 import ExperiencesPage from "../views/admin/experience/ExperiencesPage.vue";
 import RewardPage from "../views/admin/reward/RewardPage.vue";
-import RewardAddEditPage from "../views/admin/reward/RewardAddEditDialog.vue";
 import RewardRedemptionPage from "../views/admin/reward/RewardRedemptionPage.vue";
 import MaintenanceLandingPage from "../views/admin/MaintenanceLandingPage.vue";
 import MajorsPage from "../views/admin/major/MajorsPage.vue";
-import MajorAddEditPage from "../views/admin/major/MajorAddEditPage.vue";
 import Profile from "../views/Profile.vue";
 import Notifications from "../views/Notification.vue";
 
@@ -74,18 +68,6 @@ const adminTaskRoutes = [
     name: "task",
     component: TasksPage,
   },
-  // {
-  //   path: "maintenance/task/edit/:id",
-  //   name: "editTask",
-  //   component: TaskAddEditPage,
-  //   props: { isAdd: false },
-  // },
-  // {
-  //   path: "maintenance/task/add",
-  //   name: "addTask",
-  //   component: TaskAddEditPage,
-  //   props: { isAdd: true },
-  // },
 ];
 
 const adminRewardRoutes = [
@@ -93,18 +75,6 @@ const adminRewardRoutes = [
     path: "maintenance/reward",
     name: "reward",
     component: RewardPage,
-  },
-  {
-    path: "maintenance/reward/edit/:id",
-    name: "editReward",
-    component: RewardAddEditPage,
-    props: { isAdd: false },
-  },
-  {
-    path: "maintenance/reward/add",
-    name: "addReward",
-    component: RewardAddEditPage,
-    props: { isAdd: true },
   },
   {
     path: "maintenance/reward/redeem/:studentId",
@@ -119,18 +89,6 @@ const adminExperienceRoutes = [
     name: "experience",
     component: ExperiencesPage,
   },
-  {
-    path: "maintenance/experience/edit/:id",
-    name: "editExperience",
-    component: ExperienceAddEditPage,
-    props: { isAdd: false },
-  },
-  {
-    path: "maintenance/experience/add",
-    name: "addExperience",
-    component: ExperienceAddEditPage,
-    props: { isAdd: true },
-  },
 ];
 
 const adminEventRoutes = [
@@ -139,18 +97,6 @@ const adminEventRoutes = [
     name: "event",
     component: EventCardPage,
   },
-  // {
-  //   path: "maintenance/event/edit/:id",
-  //   name: "editEvent",
-  //   component: EventAddEditPage,
-  //   props: { isAdd: false },
-  // },
-  // {
-  //   path: "maintenance/event/add/:date?",
-  //   name: "addEvent",
-  //   component: EventAddEditPage,
-  //   props: { isAdd: true },
-  // },
   {
     path: "maintenance/event/attendance/:id",
     name: "attendanceEvent",
@@ -163,18 +109,6 @@ const adminBadgeRoutes = [
     path: "maintenance/badge",
     name: "badge",
     component: BadgesPage,
-  },
-  {
-    path: "maintenance/badge/edit/:id",
-    name: "editBadge",
-    component: BadgeAddEditPage,
-    props: { isAdd: false },
-  },
-  {
-    path: "maintenance/badge/add",
-    name: "addBadge",
-    component: BadgeAddEditPage,
-    props: { isAdd: true },
   },
 ];
 
@@ -196,18 +130,6 @@ const adminMajorRoutes = [
     path: "maintenance/major",
     name: "majors",
     component: MajorsPage,
-  },
-  {
-    path: "maintenance/major/edit/:id",
-    name: "editMajor",
-    component: MajorAddEditPage,
-    props: { isAdd: false },
-  },
-  {
-    path: "maintenance/major/add",
-    name: "addMajor",
-    component: MajorAddEditPage,
-    props: { isAdd: true },
   },
 ];
